refactor: /simplify — share imgProbe helper, drop dead i18n keys
- Extract the verbatim cross-origin <img> reachability probe duplicated in
  WifiHandoffOverlay and CredentialsHandoffOverlay into src/lib/handoff-probe.ts.
- Remove dead credentials.reconnecting / credentials.reconnectFailed keys
  (no t() consumer remained) from all 10 locales.

Deferred (tracked for the #167 e2e rework, where the handoff flow becomes
testable): a shared useReconnect controller hook across the 3 overlays +
SettingsApp, the CredentialsStep handoff-decision dedup, and the
getEthernetStatus/start-ap.sh nmcli micro-optimizations.
