+++ src/lib/translations.ts
@@ -123,8 +123,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "Confirm Hotspot Password",
     "credentials.reenterHotspot": "Re-enter hotspot password",
     "credentials.settingsSaved": "Settings saved! Continuing...",
-    "credentials.reconnecting": "Settings saved. Reconnecting to {url}…",
-    "credentials.reconnectFailed": "Settings were saved, but {url} is still unreachable after ~1 minute. Open the URL manually to continue setup.",
     "credentials.passwordRequired": "System password is required",
     "credentials.passwordMinLength": "System password must be at least 8 characters",
     "credentials.passwordsDontMatch": "System passwords do not match",
@@ -528,8 +526,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "Потвърдете паролата на точката за достъп",
     "credentials.reenterHotspot": "Въведете отново паролата на точката за достъп",
     "credentials.settingsSaved": "Настройките са запазени! Продължаване...",
-    "credentials.reconnecting": "Настройките са запазени. Свързване отново с {url}…",
-    "credentials.reconnectFailed": "Настройките са запазени, но {url} все още е недостъпен след около 1 минута. Отворете адреса ръчно, за да продължите настройката.",
     "credentials.passwordRequired": "Системната парола е задължителна",
     "credentials.passwordMinLength": "Системната парола трябва да е поне 8 символа",
     "credentials.passwordsDontMatch": "Системните пароли не съвпадат",
@@ -932,8 +928,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "Hotspot-Passwort bestätigen",
     "credentials.reenterHotspot": "Hotspot-Passwort erneut eingeben",
     "credentials.settingsSaved": "Einstellungen gespeichert! Weiter...",
-    "credentials.reconnecting": "Einstellungen gespeichert. Verbindung zu {url} wird wiederhergestellt…",
-    "credentials.reconnectFailed": "Einstellungen wurden gespeichert, aber {url} ist nach ~1 Minute immer noch nicht erreichbar. Öffnen Sie die URL manuell, um die Einrichtung fortzusetzen.",
     "credentials.passwordRequired": "Systempasswort ist erforderlich",
     "credentials.passwordMinLength": "Systempasswort muss mindestens 8 Zeichen lang sein",
     "credentials.passwordsDontMatch": "Systempasswörter stimmen nicht überein",
@@ -1336,8 +1330,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "Confirmar contraseña del punto de acceso",
     "credentials.reenterHotspot": "Vuelve a introducir la contraseña del punto de acceso",
     "credentials.settingsSaved": "¡Configuración guardada! Continuando...",
-    "credentials.reconnecting": "Configuración guardada. Reconectando a {url}…",
-    "credentials.reconnectFailed": "La configuración se guardó, pero {url} sigue sin estar disponible después de ~1 minuto. Abre la URL manualmente para continuar la configuración.",
     "credentials.passwordRequired": "La contraseña del sistema es obligatoria",
     "credentials.passwordMinLength": "La contraseña del sistema debe tener al menos 8 caracteres",
     "credentials.passwordsDontMatch": "Las contraseñas del sistema no coinciden",
@@ -1740,8 +1732,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "Confirmer le mot de passe du point d'accès",
     "credentials.reenterHotspot": "Ressaisissez le mot de passe du point d'accès",
     "credentials.settingsSaved": "Paramètres enregistrés ! Continuation...",
-    "credentials.reconnecting": "Paramètres enregistrés. Reconnexion à {url}…",
-    "credentials.reconnectFailed": "Les paramètres ont été enregistrés, mais {url} reste injoignable après environ 1 minute. Ouvrez l'URL manuellement pour poursuivre la configuration.",
     "credentials.passwordRequired": "Le mot de passe système est requis",
     "credentials.passwordMinLength": "Le mot de passe système doit contenir au moins 8 caractères",
     "credentials.passwordsDontMatch": "Les mots de passe système ne correspondent pas",
@@ -2144,8 +2134,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "Conferma password hotspot",
     "credentials.reenterHotspot": "Reinserisci la password dell'hotspot",
     "credentials.settingsSaved": "Impostazioni salvate! Continuazione...",
-    "credentials.reconnecting": "Impostazioni salvate. Riconnessione a {url}…",
-    "credentials.reconnectFailed": "Le impostazioni sono state salvate, ma {url} è ancora irraggiungibile dopo ~1 minuto. Apri l'URL manualmente per continuare la configurazione.",
     "credentials.passwordRequired": "La password di sistema è obbligatoria",
     "credentials.passwordMinLength": "La password di sistema deve avere almeno 8 caratteri",
     "credentials.passwordsDontMatch": "Le password di sistema non corrispondono",
@@ -2548,8 +2536,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "ホットスポットのパスワードを確認",
     "credentials.reenterHotspot": "ホットスポットのパスワードを再入力",
     "credentials.settingsSaved": "設定を保存しました！続行中...",
-    "credentials.reconnecting": "設定を保存しました。{url} に再接続しています…",
-    "credentials.reconnectFailed": "設定は保存されましたが、約1分経っても {url} に接続できません。URLを手動で開いてセットアップを続行してください。",
     "credentials.passwordRequired": "システムパスワードは必須です",
     "credentials.passwordMinLength": "システムパスワードは8文字以上にしてください",
     "credentials.passwordsDontMatch": "システムパスワードが一致しません",
@@ -2952,8 +2938,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "Hotspot-wachtwoord bevestigen",
     "credentials.reenterHotspot": "Hotspot-wachtwoord opnieuw invoeren",
     "credentials.settingsSaved": "Instellingen opgeslagen! Doorgaan...",
-    "credentials.reconnecting": "Instellingen opgeslagen. Opnieuw verbinden met {url}…",
-    "credentials.reconnectFailed": "Instellingen zijn opgeslagen, maar {url} is na ~1 minuut nog steeds onbereikbaar. Open de URL handmatig om de installatie voort te zetten.",
     "credentials.passwordRequired": "Systeemwachtwoord is verplicht",
     "credentials.passwordMinLength": "Systeemwachtwoord moet minimaal 8 tekens bevatten",
     "credentials.passwordsDontMatch": "Systeemwachtwoorden komen niet overeen",
@@ -3356,8 +3340,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "Bekräfta hotspot-lösenord",
     "credentials.reenterHotspot": "Ange hotspot-lösenordet igen",
     "credentials.settingsSaved": "Inställningar sparade! Fortsätter...",
-    "credentials.reconnecting": "Inställningar sparade. Återansluter till {url}…",
-    "credentials.reconnectFailed": "Inställningarna har sparats, men {url} är fortfarande inte tillgänglig efter ~1 minut. Öppna URL:en manuellt för att fortsätta installationen.",
     "credentials.passwordRequired": "Systemlösenord krävs",
     "credentials.passwordMinLength": "Systemlösenordet måste vara minst 8 tecken",
     "credentials.passwordsDontMatch": "Systemlösenorden matchar inte",
@@ -3760,8 +3742,6 @@ const setupTranslations: Record<Locale, Record<string, string>> = {
     "credentials.confirmHotspotPassword": "确认热点密码",
     "credentials.reenterHotspot": "重新输入热点密码",
     "credentials.settingsSaved": "设置已保存！继续中...",
-    "credentials.reconnecting": "设置已保存。正在重新连接到 {url}…",
-    "credentials.reconnectFailed": "设置已保存，但约 1 分钟后仍无法访问 {url}。请手动打开该网址以继续设置。",
     "credentials.passwordRequired": "系统密码为必填项",
     "credentials.passwordMinLength": "系统密码至少需要 8 个字符",
     "credentials.passwordsDontMatch": "两次输入的系统密码不一致",
